requirements.txt
requests
beautifulsoup4
lxml
opencage
python-dotenv